Move the offline storage logic into its own class.
